Refactor, add handler type
