+++ functions/src/index.ts
@@ -17,7 +17,7 @@
 
 import 'module-alias/register';
 import * as functions from 'firebase-functions';
-import * as cors from 'cors';
+import {onHttpsRequest} from "@/handlers";
 import {handleCreateUser} from '@/on-create-user';
 import {importCsvHandler} from '@/import-csv';
 import {importGeoJsonHandler} from '@/import-geojson';
@@ -27,22 +27,6 @@ import {
   surveyPathTemplate,
   loiPathTemplate,
 } from '@/on-write-survey';
-
-const corsOptions = {origin: true};
-const corsMiddleware = cors(corsOptions);
-
-function onHttpsRequest(handler: any) {
-  return functions.https.onRequest((req, res) =>
-    corsMiddleware(req, res, () =>
-      handler(req, res).catch((error: any) => onError(res, error))
-    )
-  );
-}
-
-function onError(res: any, err: any) {
-  console.error(err);
-  res.status(500).send('Internal error');
-}
 
 // Create user profile in database when user first logs in.
 export const onCreateUser = functions.auth.user().onCreate(handleCreateUser);
